refactor: side bar state and on open on close functions moved to App
